+++ tests/Unit/Command/ImportGridAffectationCommandTest.php
@@ -5,7 +5,7 @@ namespace App\Tests\Unit\Command;
 use App\Command\ImportGridAffectationCommand;
 use App\Entity\Territory;
 use App\Manager\TerritoryManager;
-use App\Service\GridAffectation\GridAffectationLoader;
+use App\Service\Import\GridAffectation\GridAffectationLoader;
 use App\Service\NotificationService;
 use App\Service\Parser\CsvParser;
 use App\Service\UploadHandlerService;
