+++ tests/Unit/Command/ImportSignalementCommandTest.php
@@ -6,8 +6,8 @@ use App\Command\ImportSignalementCommand;
 use App\Entity\Territory;
 use App\EventListener\ActivityListener;
 use App\Repository\TerritoryRepository;
+use App\Service\Import\Signalement\SignalementImportLoader;
 use App\Service\Parser\CsvParser;
-use App\Service\Signalement\Import\SignalementImportLoader;
 use App\Service\UploadHandlerService;
 use Doctrine\Common\EventManager;
 use Doctrine\ORM\EntityManager;
